telegram_service/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
requests==2.25.1
pika==1.2.0
python-dotenv==0.19.0
